@@ -1,25 +1,29 @@
 module github.com/Loyalsoldier/geoip
 
-go 1.18
+go 1.19
 
 require (
-	github.com/maxmind/mmdbwriter v0.0.0-20220316141846-bda76e41dd33
-	github.com/v2fly/v2ray-core/v5 v5.0.3
-	google.golang.org/protobuf v1.28.0
+	github.com/maxmind/mmdbwriter v0.0.0-20220830183856-fffdfa44ff0b
+	github.com/oschwald/maxminddb-golang v1.10.0
+	github.com/v2fly/v2ray-core/v5 v5.1.0
+	go4.org/netipx v0.0.0-20220812043211-3cc044ffd68d
+	google.golang.org/protobuf v1.28.1
 	gopkg.in/yaml.v2 v2.4.0
-	inet.af/netaddr v0.0.0-20211027220019-c74959edd3b6
 )
 
 require (
+	github.com/adrg/xdg v0.4.0 // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/lucas-clemente/quic-go v0.24.0 // indirect
-	github.com/marten-seemann/qtls-go1-17 v0.1.0 // indirect
-	github.com/oschwald/maxminddb-golang v1.7.1-0.20200819192241-1f1e288ee3f9 // indirect
-	github.com/pires/go-proxyproto v0.6.1 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	go.starlark.net v0.0.0-20211203141949-70c0e40ae128 // indirect
+	github.com/lucas-clemente/quic-go v0.28.1 // indirect
+	github.com/marten-seemann/qtls-go1-18 v0.1.2 // indirect
+	github.com/marten-seemann/qtls-go1-19 v0.1.0-beta.1 // indirect
+	github.com/pelletier/go-toml v1.9.5 // indirect
+	github.com/pires/go-proxyproto v0.6.2 // indirect
+	go.starlark.net v0.0.0-20220302181546-5411bad688d1 // indirect
 	go4.org/intern v0.0.0-20211027215823-ae77deb06f29 // indirect
-	go4.org/unsafe/assume-no-moving-gc v0.0.0-20211027215541-db492cf91b37 // indirect
-	golang.org/x/crypto v0.0.0-20211202192323-5770296d904e // indirect
-	golang.org/x/sys v0.0.0-20211205182925-97ca703d548d // indirect
+	go4.org/unsafe/assume-no-moving-gc v0.0.0-20220617031537-928513b29760 // indirect
+	golang.org/x/crypto v0.0.0-20220321153916-2c7772ba3064 // indirect
+	golang.org/x/sys v0.0.0-20220804214406-8e32c043e418 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+	inet.af/netaddr v0.0.0-20211027220019-c74959edd3b6 // indirect
 )
